Rip out util tests from test_files
